lexer refactoring
