+++ srcs/helpers/ft_char_to_str.c
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_char_to_str.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dkolida <dkolida@student.42warsaw.pl>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/07/25 12:56:16 by dkolida           #+#    #+#             */
+/*   Updated: 2024/07/25 12:56:34 by dkolida          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*char_to_str(char c)
+{
+	char	*str;
+
+	str = malloc(sizeof(char) * 2);
+	if (str == NULL)
+		return (NULL);
+	str[0] = c;
+	str[1] = '\0';
+	return (str);
+}
